wpan: refactor misc logic with util mod
