+++ embassy-stm32-wpan/src/wba/controller.rs
@@ -5,7 +5,6 @@
 
 use core::cell::RefCell;
 use core::ops::Deref;
-use core::sync::atomic::Ordering;
 
 use embassy_stm32::aes::Aes;
 use embassy_stm32::interrupt;
@@ -31,7 +30,7 @@ use crate::host_if::{MAX_BLE_PKT_SIZE, TASK_BLE_HOST_MASK, TASK_LINK_LAYER_MASK,
 use crate::linklayer_plat::{
     EVENT_CHANNEL, HARDWARE_AES, HARDWARE_PKA, HARDWARE_RNG, run_radio_high_isr, run_radio_sw_low_isr,
 };
-use crate::runner::{BLE_INIT, BLE_INIT_WAKER, BLE_SLEEPMODE_RUNNING};
+use crate::runner::{BLE_INIT, BLE_SLEEPMODE_RUNNING};
 use crate::util_seq;
 use crate::wba::ll_sys::init_ble_stack;
 
@@ -199,8 +198,7 @@ impl Controller {
         util_seq::seq_resume();
 
         // Wake the runner
-        BLE_INIT.store(true, Ordering::Release);
-        BLE_INIT_WAKER.wake();
+        BLE_INIT.set_high();
 
         Ok(Self {
             state_ptr,
